backend/requirements.txt
asgiref==3.11.0
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3